Move translator
