+++ barretenberg/cpp/src/barretenberg/stdlib/CMakeLists.txt
@@ -7,5 +7,6 @@ add_subdirectory(goblin_verifier)
 add_subdirectory(primitives)
 add_subdirectory(honk_verifier)
 add_subdirectory(transcript)
+add_subdirectory(translator_vm_verifier)
 add_subdirectory(plonk_recursion)
 add_subdirectory(protogalaxy_verifier)
