+++ barretenberg/cpp/src/barretenberg/stdlib/translator_vm_verifier/CMakeLists.txt
@@ -1,0 +1,1 @@
+barretenberg_module(stdlib_translator_vm_verifier translator_vm stdlib_honk_verifier)
