--- barretenberg/cpp/src/barretenberg/translator_vm_recursion/CMakeLists.txt
@@ -1,1 +1,0 @@
-barretenberg_module(translator_vm_recursion translator_vm stdlib_honk_verifier)
